fix(api): serve /api/environments via internal rewrite instead of a 307
The /api/environments -> /api/projects 307 redirect broke API clients that
don't follow redirects by default (httpx, Guzzle): a request that used to
return 200 got an unhandled 307. Replace the redirect with an in-process path
rewrite - when the api-environments-redirect flag is on for a team, the request
path is rewritten to /api/projects/* and routed to the projects viewset, so the
client gets a transparent 200 on the original URL with method, body, query
string, and auth intact.

Same per-team flag gating and Deprecation/Sunset/Link headers; environment-only
routes still pass through to the legacy route. Removes the now-dead 307 response
class and the test helper that followed the redirect.
